feat(targeting): canonicalize seller_agent_url via shared urlcanon
seller_agent_url selects which seller's active package set the
context-agent returns. The read path keyed on the raw request string,
but the AdCP spec compares URL identifiers under URL-identifier
canonicalization (docs/reference/url-canonicalization), not
byte-equality, so a request carrying
"https://Seller.Example.com:443/agent" missed a set registered under
"https://seller.example.com/agent".

The exact algorithm already existed as the private canonicalTargetURI in
adcp/signing. Extract it to a new stdlib-only leaf package
adcp/urlcanon as the exported Canonicalize so a single implementation
is shared by both the signing @target-uri path and the context-agent
read path — no split-brain. adcp/signing now delegates to it
(canonicalTargetURI and canonicalAuthority call urlcanon); behavior and
the conformance vectors are unchanged.

targeting/engine.go canonicalizes req.SellerAgentURL once, the single
read-path chokepoint, before ActivePackages. An empty value passes
through unchanged; a non-empty malformed value fails safe to an empty
active set (empty offers) and logs once — never 5xx, never a raw lookup.

Wire root -> adcp: root go.mod requires/replaces adcp, and every module
that replaces the root module gains its own replace for adcp. adcp
declares go 1.26.2, so root and the dependent modules bump to match.
