+++ cmd/router/go.mod
@@ -1,6 +1,6 @@
 module github.com/adcontextprotocol/adcp-go/cmd/router
 
-go 1.25.0
+go 1.26.2
 
 require (
 	github.com/adcontextprotocol/adcp-go v0.0.0
@@ -8,6 +8,7 @@ require (
 )
 
 require (
+	github.com/adcontextprotocol/adcp-go/adcp v0.0.0 // indirect
 	github.com/hashicorp/golang-lru/v2 v2.0.7 // indirect
 	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
 	golang.org/x/crypto v0.51.0 // indirect
@@ -19,3 +20,5 @@ replace (
 	github.com/adcontextprotocol/adcp-go => ../../
 	github.com/adcontextprotocol/adcp-go/targeting/prommetrics => ../../targeting/prommetrics
 )
+
+replace github.com/adcontextprotocol/adcp-go/adcp => ../../adcp
