+++ reference/context-agent/go.mod
@@ -1,6 +1,6 @@
 module github.com/adcontextprotocol/adcp-go/reference/context-agent
 
-go 1.25.0
+go 1.26.2
 
 require (
 	github.com/adcontextprotocol/adcp-go v0.0.0
@@ -9,6 +9,7 @@ require (
 )
 
 require (
+	github.com/adcontextprotocol/adcp-go/adcp v0.0.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/hashicorp/golang-lru/v2 v2.0.7 // indirect
 	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
@@ -23,3 +24,5 @@ replace (
 	github.com/adcontextprotocol/adcp-go => ../../
 	github.com/adcontextprotocol/adcp-go/targeting/prommetrics => ../../targeting/prommetrics
 )
+
+replace github.com/adcontextprotocol/adcp-go/adcp => ../../adcp
