+++ targeting/prommetrics/go.mod
@@ -1,10 +1,11 @@
 module github.com/adcontextprotocol/adcp-go/targeting/prommetrics
 
-go 1.25.0
+go 1.26.2
 
 require github.com/adcontextprotocol/adcp-go v0.0.0
 
 require (
+	github.com/adcontextprotocol/adcp-go/adcp v0.0.0 // indirect
 	github.com/hashicorp/golang-lru/v2 v2.0.7 // indirect
 	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
 	golang.org/x/crypto v0.51.0 // indirect
@@ -20,3 +21,5 @@ require (
 )
 
 replace github.com/adcontextprotocol/adcp-go => ../../
+
+replace github.com/adcontextprotocol/adcp-go/adcp => ../../adcp
